move cli into pkg/cmd
