+++ pkg/cmd/http/http.go
@@ -1,7 +1,7 @@
 package http
 
 import (
-	rootcmd "github.com/solo-io/gloo-connect/pkg/cmd"
+	"github.com/solo-io/gloo-connect/pkg/cmd/gloo_client"
 	"github.com/solo-io/gloo-connect/pkg/runner"
 	"github.com/solo-io/gloo/pkg/bootstrap/configstorage"
 	"github.com/spf13/cobra"
@@ -18,7 +18,7 @@ func Cmd(rc *runner.RunConfig) *cobra.Command {
 				return err
 			}
 
-			gc := rootcmd.GlooClient{Store: store}
+			gc := gloo_client.GlooClient{Store: store}
 
 			return gc.Demo()
 		},
